requirements.txt
# Core dependencies
typer>=0.9.0
jinja2>=3.1.2
weasyprint>=62.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.1
certifi>=2024.6.2

# Malware analysis helpers
r2pipe>=1.9.0
pefile>=2023.2.7
python-magic>=0.4.27
pyzipper>=0.3.6
yara-python>=4.3.0
capstone>=5.0.0
numpy>=1.26.0

# Machine learning
scikit-learn>=1.4.0
joblib>=1.3.0

# Optional but recommended
ipaddress; python_version < "3.3"   # builtin da Python 3.3 in poi

# Development & formatting
black>=24.4.2
flake8>=7.1.0
mypy>=1.10.0

#host checker
tranco
tldextract